fix: always use google.cloud.storage library to download from GCS (#212)
We implemented this as a workaround for an issue when downloading large files via HTTPS but we should consider keeping it in the long run if it makes our code more reliable.

The PR also restructures the download_or_copy_file function a little bit to make it more readable but without making any functional changes (except for the URL rewriting at the top).