Simplify video concatenation (#217)
- Removed calls to the compositevideo and useless calls to ML Gateway
Factory. This was ackward as we had to instanciate factories not to
really leverage them, as well as for the prompt, and do this in test
mode
- directly had to change a ffmpegwrapper function to concatenate videos
as it needed manipulating files instead of list of videos, preferred
doing this at lowest possible level
- now concatenate_videos function supports a list of videos

Cherry-picked and modified commit 854189b from the monorepo.

---------

Co-authored-by: JF Mac <jf@vikit,ai>